requirements.txt
numpy>=1.20.0
opencv-python>=4.5.5
torch>=1.12.0
torchvision>=0.13.0
scikit-image>=0.19.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
pillow>=9.0.0
tqdm>=4.62.0
scipy>=1.7.0
mediapipe>=0.8.10
dlib>=19.22.0
onnxruntime>=1.10.0
segmentation-models-pytorch>=0.3.0
albumentations>=1.1.0
kornia>=0.6.0
einops>=0.4.1
transformers>=4.18.0
diffusers>=0.12.0
accelerate>=0.16.0
huggingface-hub>=0.12.0
requests>=2.27.0
python-dotenv>=0.20.0
fastapi>=0.85.0
uvicorn>=0.18.0
python-multipart>=0.0.5 